Rename hook to fetchPortfolioBalances action func since hook is deprecated with Portfolio store addition
